Update you-should-use.plugin.zsh
This *.plugin.zsh-file should be the the init-file according to zsh-plugin-standards. The plugin itself will be the *.zsh-file. In this case the init-file will be name you-should-use.plugin.zsh and the plugin-file will be you-should-use.zsh
Furthermore the $0 handling should be somehing like this:
# Standardized $0 handling (See https://zdharma-continuum.github.io/Zsh-100-Commits-Club/Zsh-Plugin-Standard.html)
`0="${ZERO:-${${0:#$ZSH_ARGZERO}:-${(%):-%N}}}"`

or set $0 with use of %x prompt expansion

`0="${ZERO:-${${${(M)${0::=${(%):-%x}}:#/*}:-$PWD/$0}:A}}"`

I use to set both since it doesn't hurt. Just in case someone asks for one or the other.

Then one more line should be added:
local ZERO="$0"